avoid using set for norm columns deduplication

+++ core/analysis/text_token_stream.cpp
@@ -1027,7 +1027,7 @@ bool text_token_stream::reset(const string_ref& data) {
     data_utf8_ref = data_utf8;
   }
 
-  if (data_utf8_ref.size() > std::numeric_limits<int32_t>::max()) {
+  if (data_utf8_ref.size() > static_cast<uint32_t>(std::numeric_limits<int32_t>::max())) {
     return false; // ICU UnicodeString signatures can handle at most INT32_MAX
   }
